OpenFlow: use same timer throughout all component lifetime
Calls to the modified methods (all configuration changes through the
CLI, for example) of OpenFlowDeviceProvider and OpenFlowRuleProvider
cause a switch from executor to another one on PortStatsCollector and
FlowStats collector respectively.

Since these Collectors are constructed with a Timer argument, we might
as well use those timers properly right away.

Change-Id: I32aa8c89d9615ea065e5fd9fd7dfc70da96b9739